+++ semgrep-core/src/core/ast/AST_generic.ml
@@ -758,9 +758,9 @@ and concat_string_kind =
    * and some semgrep users may want to explicitely match only f-strings,
    * which is why we record this information here.
    * update: also use for interpolated Scala strings
-   * TODO: add of string ('f' or something else)
+   * update: add string argument to support arbitary string interpolaters in scala
    *)
-  | FString
+  | FString of string
   (* Javascript uses a special syntax called tagged template literals, e.g.,
    * foo`template string = ${id}`. We must use a different representation
    * for foo(`template string = ${id}`) because some semgrep users want